Update TARDISSaveLocationCommand.java
Attempt to update the /tardis save [name] command to require use of a Disk in Hard mode.  If I'm reading the code right, this should warn the player to use a disk if they attempt to use the /tardis save command on Hard mode with no disk in their hand.